requirements.txt
alabaster==0.7.8
amqp==5.0.2
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.1
attrs==20.3.0
awesome-slugify==1.6.5
Babel==2.6.0
backcall==0.2.0
bcrypt==3.1.7
billiard==3.6.3.0
black==20.8b1
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
celery==5.0.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
ddtrace==0.44.0
debtcollector==2.2.0
decorator==4.4.2
defer==1.0.6
distlib==0.3.0
distro==1.6.0
distro-info===0.23ubuntu1
Django==3.1.4
django-behaviors==0.5.1
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
django-healthchecks==1.4.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-redis==5.0.0
django-rest-swagger==2.2.0
django-storage-swift==1.2.19
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docker==5.0.2
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
drf-nested-routers==0.92.5
drf-yasg==1.20.0
duplicity==0.8.12.0
entrypoints==0.3
environ==1.0
eradicate==2.0.0
Faker==0.9.1
fasteners==0.14.1
feedparser==5.2.1
filelock==3.0.12
flake8==3.8.4
flake8-black==0.2.1
flake8-bugbear==20.11.1
flake8-eradicate==1.0.0
flake8-isort==4.0.0
flake8-print==4.0.0
freezegun==1.0.0
future==0.18.2
gitsome==0.8.0
httplib2==0.14.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
inflection==0.5.1
iniconfig==1.1.1
intervaltree==3.1.0
ipython==7.19.0
ipython-genutils==0.2.0
iso8601==0.1.13
isort==5.6.4
itypes==1.2.0
jedi==0.17.2
Jinja2==2.10.1
jsonschema==3.2.0
keyring==18.0.1
keystoneauth1==4.3.0
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
mixer==6.1.3
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.0
mypy-extensions==0.4.3
netaddr==0.8.0
netifaces==0.10.9
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
os-service-types==1.7.0
oslo.config==8.4.0
oslo.i18n==5.0.1
oslo.serialization==4.0.1
oslo.utils==4.7.0
packaging==20.7
paramiko==2.6.0
parso==0.7.1
pathspec==0.8.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pip-tools==5.4.0
pipenv==11.9.0
pluggy==0.13.1
ply==3.11
prompt-toolkit==3.0.8
protobuf==3.14.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pycairo==1.16.2
pycodestyle==2.6.0
pycups==1.9.73
pyflakes==2.2.0
Pygments==2.7.3
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.0
pytest==6.1.2
pytest-django==4.1.0
pytest-env==0.6.2
pytest-freezegun==0.4.2
pytest-mock==3.3.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.19.0
python-keystoneclient==4.2.0
python-magic==0.4.18
python-swiftclient==3.11.0
pytz==2020.4
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
regex==2020.11.13
reportlab==3.5.34
requests==2.22.0
requests-mock==1.8.0
requests-unixsocket==0.2.0
rfc3986==1.4.0
roman==2.0.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
sentry-sdk==0.19.5
setproctitle==1.1.10
simplejson==3.16.0
six==1.15.0
sortedcontainers==2.3.0
Sphinx==1.8.5
sqlparse==0.4.1
statsd==3.3.0
stevedore==3.3.0
systemd-python==234
tenacity==6.2.0
testfixtures==6.16.0
text-unidecode==1.2
texttable==1.6.4
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==0.4.21
uritemplate==3.0.1
urllib3==1.25.11
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
wcwidth==0.2.5
websocket-client==0.59.0
wrapt==1.12.1
xkit==0.0.0
xonsh==0.9.13
zipp==1.0.0
